+++ lib/invoca/metrics.rb
@@ -4,6 +4,7 @@ require 'active_support'
 require 'active_support/core_ext'
 
 require "invoca/metrics/version"
+require "invoca/metrics/statsd_client"
 require "invoca/metrics/client"
 require "invoca/metrics/direct_metric"
 require "invoca/metrics/batch"
@@ -58,17 +59,24 @@ module Invoca
       extend ActiveSupport::Concern
 
       module ClassMethods
-        def metrics
-          @metrics ||= metrics_for(config_key: Invoca::Metrics.default_config_key)
+        @metrics_namespace = nil
+
+        def metrics_namespace(namespace)
+          @metrics_namespace = namespace
         end
 
-        def metrics_for(config_key:)
-          @metrics_for ||= {}
-          @metrics_for[config_key] ||=
-            begin
-              metrics_config = Invoca::Metrics.config[config_key] || {}
-              Client.metrics(Invoca::Metrics.default_client_config.merge(metrics_config))
-            end
+        def metrics
+          metrics_for(config_key: Invoca::Metrics.default_config_key)
+        end
+
+        def metrics_for(config_key:, namespace: nil)
+          config_from_key      = Invoca::Metrics.config[config_key] || {}
+          metrics_config       = if (effective_namespace = namespace || @metrics_namespace)
+                                   config_from_key.merge(namespace: effective_namespace)
+                                 else
+                                   config_from_key
+                                 end
+          Client.metrics(Invoca::Metrics.default_client_config.merge(metrics_config))
         end
       end
 
@@ -76,8 +84,8 @@ module Invoca
         self.class.metrics
       end
 
-      def metrics_for(config_key:)
-        self.class.metrics_for(config_key: config_key)
+      def metrics_for(config_key:, namespace: nil)
+        self.class.metrics_for(config_key: config_key, namespace: namespace)
       end
     end
   end